Revert "Turn off logging for containers for some classes"
