+++ config-model/src/main/java/com/yahoo/vespa/model/admin/LogserverContainerCluster.java
@@ -25,6 +25,9 @@ public class LogserverContainerCluster extends ContainerCluster<LogserverContain
     }
 
     @Override
+    protected void doPrepare(DeployState deployState) { }
+
+    @Override
     public void getConfig(QrStartConfig.Builder builder) {
         super.getConfig(builder);
         builder.jvm.heapsize(128);
